careerpilot-ai/python-service/run.py
"""Entry point for the CareerPilot AI Python microservice."""
import os
from dotenv import load_dotenv
from app import create_app

load_dotenv()

app = create_app()

if __name__ == '__main__':
    port = int(os.getenv('PORT', 8000))
    debug = os.getenv('FLASK_ENV') == 'development'
    print(f'[CareerPilot] Python service running on http://localhost:{port}')
    app.run(host='0.0.0.0', port=port, debug=debug)
